ISS: disable the deletion behavior on debug builds
this should help surface vm data unlock issues

Test: manually ran KEK test case to make sure appsearch doesn't recover
from the issue
Bug: 444419346
Flag: com.android.appsearch.flags.enable_isolated_storage

+++ service/java/com/android/server/appsearch/isolated_storage_service/IsolatedStorageServiceManager.java
@@ -100,6 +100,9 @@ public class IsolatedStorageServiceManager {
     private static final long VM_STATUS_CHECK_INTERVAL_SECONDS = 60; // 1 minute
 
     private static final UserHandle ISOLATED_STORAGE_USER = UserHandle.SYSTEM;
+    private static final String SYSTEM_PROPERTY_ENABLE_DEBUG_BUILD = "ro.debuggable";
+    private static final boolean IS_DEBUG_BUILD =
+            SystemProperties.getInt(SYSTEM_PROPERTY_ENABLE_DEBUG_BUILD, /* def= */ 0) == 1;
 
     private final Context mContext;
     private final ServiceAppSearchConfig mAppSearchConfig;
@@ -986,12 +989,21 @@ public class IsolatedStorageServiceManager {
                 mConsecutiveVmDataUnlockDoeErrorsLocked++;
                 if (mConsecutiveVmDataUnlockDoeErrorsLocked
                         >= CONSECUTIVE_VM_DATA_UNLOCK_DOE_ERROR_THRESHOLD) {
-                    if (mIsolatedStorageService.deleteVm()) {
-                        mConsecutiveVmDataUnlockDoeErrorsLocked = 0;
+                    // keep the vm on debug builds to surface the unlock issues
+                    if (IS_DEBUG_BUILD) {
                         Log.wtf(
                                 TAG,
-                                "Deleted the VM due to consecutive DOE errors when trying to"
-                                        + " unlock vm data");
+                                "Saw at least "
+                                        + CONSECUTIVE_VM_DATA_UNLOCK_DOE_ERROR_THRESHOLD
+                                        + " consecutive DOE errors when trying to unlock vm data");
+                    } else {
+                        if (mIsolatedStorageService.deleteVm()) {
+                            mConsecutiveVmDataUnlockDoeErrorsLocked = 0;
+                            Log.wtf(
+                                    TAG,
+                                    "Deleted the VM due to consecutive DOE errors when trying to"
+                                            + " unlock vm data");
+                        }
                     }
                 }
                 throw e;
